Conflicts:
	src/com/java/dao/ImplementUserRepository.java

+++ src/com/java/dto/Transaction.java
@@ -19,5 +19,5 @@ import lombok.ToString;
 public class Transaction {
 	private int sender;
 	private int receiver;
-	private double nominal;
+	private int nominal;
 }
